Split off init mod from plugin
