+++ src/main.rs
@@ -15,6 +15,7 @@ extern crate regex;
 extern crate "rustc-serialize" as serialize;
 extern crate uuid;
 
+mod init;
 mod plugin;
 
 fn handle_stream(mut stream: UnixStream) {
@@ -25,7 +26,7 @@ fn main() {
     let path = format!("/tmp/{}.sock", Uuid::new_v4().to_simple_string());
     let sock = UnixListener::bind(path.as_slice()).unwrap();
 
-    let mut plugins = plugin::read_init(&Path::new("init")).unwrap();
+    let mut plugins = init::read_init(&Path::new("init")).unwrap();
     for p in plugins.iter_mut() {
         if let Err(ref err) = p.load(path.as_slice()) {
             println!("{}", err);
